+++ src/languageFeatures/internal/brunoFiles/completionItems/provideCodeBlocksCompletionItems.ts
@@ -14,6 +14,7 @@ import {
     RequestFileBlockName,
     OutputChannelLogger,
     mapFromVsCodePosition,
+    Block,
 } from "../../../../shared";
 import { getCodeBlocks } from "../shared/codeBlocksUtils/getCodeBlocks";
 import { getPositionWithinTempJsFile } from "../shared/codeBlocksUtils/getPositionWithinTempJsFile";
@@ -21,6 +22,14 @@ import { mapToRangeWithinBruFile } from "../shared/codeBlocksUtils/mapToRangeWit
 import { getRequestFileDocumentSelector } from "../shared/getRequestFileDocumentSelector";
 import { waitForTempJsFileToBeInSyncWithBruFile } from "../shared/codeBlocksUtils/waitForTempJsFileToBeInSyncWithBruFile";
 import { TempJsFileUpdateQueue } from "../../shared/temporaryJsFilesUpdates/external/tempJsFileUpdateQueue";
+
+type CompletionItemRange =
+    | VsCodeRange
+    | {
+          inserting: VsCodeRange;
+          replacing: VsCodeRange;
+      }
+    | undefined;
 
 export function provideCodeBlocksCompletionItems(
     queue: TempJsFileUpdateQueue,
@@ -83,6 +92,8 @@ export function provideCodeBlocksCompletionItems(
                     return undefined;
                 }
 
+                const startTimeForFetchingCompletions = performance.now();
+
                 const resultFromJsFile =
                     await commands.executeCommand<CompletionList>(
                         "vscode.executeCompletionItemProvider",
@@ -98,51 +109,233 @@ export function provideCodeBlocksCompletionItems(
                         ),
                     );
 
-                return new CompletionList<CompletionItem>(
-                    resultFromJsFile.items.map((item) => ({
-                        ...item,
-                        /* Without unsetting the command field, selecting an exported function causes the `require` statement 
-                        to be inserted in the temp js file. */
-                        command: undefined,
-                        range: item.range
-                            ? item.range instanceof VsCodeRange
-                                ? (mapToRangeWithinBruFile(
-                                      blocksToCheck,
-                                      temporaryJsDoc.getText(),
-                                      item.range,
-                                      logger,
-                                  ) as VsCodeRange)
-                                : {
-                                      inserting: mapToRangeWithinBruFile(
-                                          blocksToCheck,
-                                          temporaryJsDoc.getText(),
-                                          item.range.inserting,
-                                          logger,
-                                      ) as VsCodeRange,
-                                      replacing: mapToRangeWithinBruFile(
-                                          blocksToCheck,
-                                          temporaryJsDoc.getText(),
-                                          item.range.replacing,
-                                          logger,
-                                      ) as VsCodeRange,
-                                  }
-                            : undefined,
-                        textEdit: item.textEdit
-                            ? new TextEdit(
-                                  mapToRangeWithinBruFile(
-                                      blocksToCheck,
-                                      temporaryJsDoc.getText(),
-                                      item.textEdit.range,
-                                  ) as VsCodeRange,
-                                  item.textEdit.newText,
-                              )
-                            : undefined,
-                    })),
+                const endTimeForFetchingCompletions = performance.now();
+                logger?.trace(
+                    `Fetching completion items from temp JS file duration: ${
+                        endTimeForFetchingCompletions -
+                        startTimeForFetchingCompletions
+                    } ms`,
+                );
+
+                if (token.isCancellationRequested) {
+                    logger?.debug(
+                        `Cancellation requested for completion provider for code blocks while fetching completons from temp JS file.`,
+                    );
+                    return undefined;
+                }
+
+                const startTimeForMappingCompletions = performance.now();
+
+                const currentTempJsContent = temporaryJsDoc.getText();
+
+                const knownRangeMappings: {
+                    rangeInTempJsFile: VsCodeRange;
+                    rangeInBruFile: VsCodeRange;
+                }[] = [];
+
+                const result = new CompletionList<CompletionItem>(
+                    resultFromJsFile.items
+                        .map((item) =>
+                            getMappedItem(
+                                item,
+                                blockInBruFile,
+                                currentTempJsContent,
+                                knownRangeMappings,
+                                logger,
+                            ),
+                        )
+                        .filter((mappedItem) => mappedItem != undefined),
                     resultFromJsFile.isIncomplete,
                 );
+
+                const endTimeForMappingCompletions = performance.now();
+                logger?.trace(
+                    `Mapping completion items from temp JS file to bru file duration: ${
+                        endTimeForMappingCompletions -
+                        startTimeForMappingCompletions
+                    } ms`,
+                );
+
+                if (result.items.length < resultFromJsFile.items.length) {
+                    logger?.debug(
+                        `Only managed to map ${result.items.length} / ${resultFromJsFile.items.length} completion items from temp JS file.`,
+                    );
+                }
+
+                return result;
             },
         },
         ".",
         "/",
     );
+}
+
+/**
+ * Checks if all range related fields can be mapped and if not, undefined is returned. Otherwise, the mapped item is returned.
+ */
+function getMappedItem(
+    item: CompletionItem,
+    blockInBruFile: Block,
+    currentTempJsContent: string,
+    knownRangeMappings: {
+        rangeInTempJsFile: VsCodeRange;
+        rangeInBruFile: VsCodeRange;
+    }[],
+    logger?: OutputChannelLogger,
+): CompletionItem | undefined {
+    const { range: mappedRange, couldBeMapped: couldRangeBeMapped } =
+        getMappedItemRange(
+            item.range,
+            blockInBruFile,
+            currentTempJsContent,
+            knownRangeMappings,
+            logger,
+        );
+
+    if (!couldRangeBeMapped) {
+        return undefined;
+    }
+
+    const { textEdit: mappedTextEdit, couldBeMapped: couldTextEditBeMapped } =
+        getMappedItemTextEdit(
+            item.textEdit,
+            blockInBruFile,
+            currentTempJsContent,
+            knownRangeMappings,
+            logger,
+        );
+
+    if (!couldTextEditBeMapped) {
+        return undefined;
+    }
+    return {
+        ...item,
+        /* Without unsetting the command field, selecting an exported function causes the `require` statement 
+                        to be inserted in the temp js file. */
+        command: undefined,
+        range: mappedRange,
+        textEdit: mappedTextEdit,
+    };
+}
+
+function getMappedItemRange(
+    itemRange: CompletionItemRange,
+    blockInBruFile: Block,
+    currentTempJsContent: string,
+    knownRangeMappings: {
+        rangeInTempJsFile: VsCodeRange;
+        rangeInBruFile: VsCodeRange;
+    }[],
+    logger?: OutputChannelLogger,
+): { range: CompletionItemRange; couldBeMapped: boolean } {
+    if (itemRange == undefined) {
+        return { range: undefined, couldBeMapped: true };
+    }
+
+    if (itemRange instanceof VsCodeRange) {
+        return {
+            range: mapTempJsRangeToBruFileRange(
+                blockInBruFile,
+                currentTempJsContent,
+                itemRange,
+                knownRangeMappings,
+                logger,
+            ),
+            couldBeMapped: true,
+        };
+    }
+
+    const mappedInsertion = mapTempJsRangeToBruFileRange(
+        blockInBruFile,
+        currentTempJsContent,
+        itemRange.inserting,
+        knownRangeMappings,
+        logger,
+    );
+
+    const mappedReplacement = mapTempJsRangeToBruFileRange(
+        blockInBruFile,
+        currentTempJsContent,
+        itemRange.replacing,
+        knownRangeMappings,
+        logger,
+    );
+
+    if (!mappedInsertion || !mappedReplacement) {
+        return { range: undefined, couldBeMapped: false };
+    }
+
+    return {
+        range: {
+            inserting: mappedInsertion,
+            replacing: mappedReplacement,
+        },
+        couldBeMapped: true,
+    };
+}
+
+function getMappedItemTextEdit(
+    itemTextEdit: TextEdit | undefined,
+    blockInBruFile: Block,
+    currentTempJsContent: string,
+    knownRangeMappings: {
+        rangeInTempJsFile: VsCodeRange;
+        rangeInBruFile: VsCodeRange;
+    }[],
+    logger?: OutputChannelLogger,
+): { textEdit: TextEdit | undefined; couldBeMapped: boolean } {
+    if (!itemTextEdit) {
+        return { textEdit: undefined, couldBeMapped: true };
+    }
+
+    const mappedTextEditRange = mapTempJsRangeToBruFileRange(
+        blockInBruFile,
+        currentTempJsContent,
+        itemTextEdit.range,
+        knownRangeMappings,
+        logger,
+    );
+
+    if (!mappedTextEditRange) {
+        return { textEdit: undefined, couldBeMapped: false };
+    }
+
+    return {
+        textEdit: new TextEdit(mappedTextEditRange, itemTextEdit.newText),
+        couldBeMapped: true,
+    };
+}
+
+function mapTempJsRangeToBruFileRange(
+    blockInBruFile: Block,
+    fullJsFileContent: string,
+    rangeInJsFile: VsCodeRange,
+    knownRangeMappings: {
+        rangeInTempJsFile: VsCodeRange;
+        rangeInBruFile: VsCodeRange | undefined;
+    }[],
+    logger?: OutputChannelLogger,
+) {
+    const knownMapping = knownRangeMappings.find(
+        ({ rangeInTempJsFile: rangeWithKnownMapping }) =>
+            rangeInJsFile.isEqual(rangeWithKnownMapping),
+    );
+
+    if (knownMapping) {
+        return knownMapping.rangeInBruFile;
+    }
+
+    const mappedRange = mapToRangeWithinBruFile(
+        blockInBruFile,
+        fullJsFileContent,
+        rangeInJsFile,
+        logger,
+    );
+
+    knownRangeMappings.push({
+        rangeInTempJsFile: rangeInJsFile,
+        rangeInBruFile: mappedRange,
+    });
+
+    return mappedRange;
 }